backend/requirements.txt
colour-science>=0.4.6
matplotlib>=3.5
numpy>=1.21
flask>=2.3
werkzeug>=2.3
flask-cors>=4.0 